-UI elements added programatically -Base FSM implemented

+++ BraggartPort/deck.swift
@@ -14,8 +14,8 @@ class Card : SKSpriteNode
 {
     let cardname:String
     let cardtexture:SKTexture
-    let boastcardvalue:Int8
-    let finalcardvalue:Int8
+    let boastcardvalue:Int
+    let finalcardvalue:Int
     var cardnames:[String] = ["UnfortunateTrout","AngryChicken","ViciousCat","Johan","TerrifiedUrchin","DrunkenDwarves",
     "SimonTheOgre","AnnoyingBard","TribeOfFaeries","CaptainCarlo","ForestDenizens", "Wizards", "Ghoul", "Damned Spirits","PompousCultists",
     "HairyRaiders","HighPriest","UgKrugGlug","Gladiator","SkilledAssassins","Mercenaries", "AmazonWarrior", "RogueMagician", "FellKnight", "King", "UndeadKing",
@@ -37,8 +37,8 @@ class Card : SKSpriteNode
     "SceneBedchamberCard", "SceneTempleCard", "SceneSorceryCard", "SceneBootsSmileCard", "SceneBattlefieldCard", "SceneWarlordCard", "SceneFrozenCard", "SceneCitadelCard", "ResultCryingCard",
     "ResultBannedCard", "ResultSmellCard", "ResultBetterPersonCard", "ResultGlowCard", "ResultBarmaidsCard", "ResultCourtCard", "ResultDeedsCard", "ResultKnightedCard", "ResultPaintingCard",
     "ResultTreasuresCard", "ResultArmiesCard", "ResultChampionCard", "ResultGodofwarCard"]
-    var boastcardvalues:[Int8] = [1,1,2,2,2,3,3,3,3,4,4,4,4,5,5,5,5,5,6,6,6,6,6,7,7,7,7,8,8,8,9,9,1,1,2,2,2,3,3,3,3,4,4,4,4,5,5,5,5,5,6,6,6,6,6,7,7,7,7,8,8,8,9,9,1,2,2,3,3,4,4,5,5,6,6,7,7,8,1,2,2,3,3,4,4,5,5,6,6,7,7,8]
-    var finalcardvalues:[Int8] = [1,1,1,1,1,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3,3,4,4,4,4,4,4,4,5,5,1,1,1,1,1,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3,3,4,4,4,4,4,4,4,5,5,1,1,1,2,2,2,2,3,3,3,3,4,4,4,1,1,1,2,2,2,2,3,3,3,3,4,4,4]
+    var boastcardvalues:[Int] = [1,1,2,2,2,3,3,3,3,4,4,4,4,5,5,5,5,5,6,6,6,6,6,7,7,7,7,8,8,8,9,9,1,1,2,2,2,3,3,3,3,4,4,4,4,5,5,5,5,5,6,6,6,6,6,7,7,7,7,8,8,8,9,9,1,2,2,3,3,4,4,5,5,6,6,7,7,8,1,2,2,3,3,4,4,5,5,6,6,7,7,8]
+    var finalcardvalues:[Int] = [1,1,1,1,1,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3,3,4,4,4,4,4,4,4,5,5,1,1,1,1,1,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3,3,4,4,4,4,4,4,4,5,5,1,1,1,2,2,2,2,3,3,3,3,4,4,4,1,1,1,2,2,2,2,3,3,3,3,4,4,4]
     required init?(coder aDecoder: NSCoder)
     {
         fatalError("NSCoding not supported")
@@ -52,11 +52,11 @@ class Card : SKSpriteNode
         finalcardvalue = finalcardvalues[index]
         super.init(texture: cardtexture, color: .clear, size: CGSize(width:120, height:180))
     }
-    func getBoastValue()->Int8
+    func getBoastValue()->Int
     {
         return boastcardvalue
     }
-    func getFinalValue()->Int8
+    func getFinalValue()->Int
     {
         return finalcardvalue
     }
